services/barilo-rest-api/pyproject.toml
[tool.poetry]
name = "barilo-rest-api"
version = "0.1.0"
description = ""
authors = ["Pedro Santos <pedro357bm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
ipython = "^8.24.0"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"
ruff = "^0.4.5"
pandas = "^2.2.2"
structlog = "^24.1.0"
django-cors-headers = "^4.3.1"
pika = "^1.3.2"
grpcio = "^1.64.0"
grpcio-tools = "^1.64.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
